Code Refactoring and UI creation
