+++ app/build.gradle
@@ -7,7 +7,7 @@ apply plugin: 'kotlin-android-extensions'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.kata.ticktactoe"
+        applicationId "com.kata.tictactoe"
         minSdkVersion 23
         targetSdkVersion 28
         versionCode 1
