+++ app/src/main/java/com/kata/tictactoe/MainActivity.kt
@@ -1,7 +1,7 @@
-package com.kata.ticktactoe
+package com.kata.tictactoe
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 
 class MainActivity : AppCompatActivity() {
