+++ app/src/main/res/layout/activity_main.xml
@@ -1,19 +1,125 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
+<TableLayout
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:tools="http://schemas.android.com/tools"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        android:padding="@dimen/padding_5dp"
+        android:id="@+id/tableLayout"
+        tools:context="com.kata.tictactoe.MainActivity">
 
-    <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Hello World!"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="parent"/>
+    <TableRow
+            android:layout_width="match_parent"
+            android:layout_height="@dimen/row_height_0dp"
+            android:layout_weight="1">
+        <Button
+                android:id="@+id/button1"
+                android:background="@color/buttonColor"
+                android:layout_margin="@dimen/padding_5dp"
+                style="?android:attr/buttonBarButtonStyle"
+                android:layout_width="@dimen/row_item_width_0dp"
+                android:layout_weight="1"
+                android:layout_height="match_parent"/>
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+        <Button
+                android:id="@+id/button2"
+                style="?android:attr/buttonBarButtonStyle"
+                android:background="@color/buttonColor"
+                android:layout_margin="@dimen/padding_5dp"
+                android:layout_width="@dimen/row_item_width_0dp"
+                android:layout_weight="1"
+                android:layout_height="match_parent"/>
+        <Button
+                android:id="@+id/button3"
+                style="?android:attr/buttonBarButtonStyle"
+                android:background="@color/buttonColor"
+                android:layout_margin="@dimen/padding_5dp"
+                android:layout_width="@dimen/row_item_width_0dp"
+                android:layout_weight="1"
+                android:layout_height="match_parent"/>
+    </TableRow>
+
+    <TableRow
+            android:layout_width="match_parent"
+            android:layout_height="@dimen/row_height_0dp"
+            android:layout_weight="1">
+        <Button
+                android:id="@+id/button4"
+                style="?android:attr/buttonBarButtonStyle"
+                android:background="@color/buttonColor"
+                android:layout_margin="@dimen/padding_5dp"
+                android:layout_width="@dimen/row_item_width_0dp"
+                android:layout_weight="1"
+                android:layout_height="match_parent"/>
+        <Button
+                android:id="@+id/button5"
+                style="?android:attr/buttonBarButtonStyle"
+                android:background="@color/buttonColor"
+                android:layout_margin="@dimen/padding_5dp"
+                android:layout_width="@dimen/row_item_width_0dp"
+                android:layout_weight="1"
+                android:layout_height="match_parent"/>
+        <Button
+                android:id="@+id/button6"
+                style="?android:attr/buttonBarButtonStyle"
+                android:background="@color/buttonColor"
+                android:layout_margin="@dimen/padding_5dp"
+                android:layout_width="@dimen/row_item_width_0dp"
+                android:layout_weight="1"
+                android:layout_height="match_parent"/>
+    </TableRow>
+
+    <TableRow
+            android:layout_width="match_parent"
+            android:layout_height="@dimen/row_height_0dp"
+            android:layout_weight="1">
+        <Button
+                android:id="@+id/button7"
+                style="?android:attr/buttonBarButtonStyle"
+                android:background="@color/buttonColor"
+                android:layout_margin="@dimen/padding_5dp"
+                android:layout_width="@dimen/row_item_width_0dp"
+                android:layout_weight="1"
+                android:layout_height="match_parent"/>
+        <Button
+                android:id="@+id/button8"
+                style="?android:attr/buttonBarButtonStyle"
+                android:background="@color/buttonColor"
+                android:layout_margin="@dimen/padding_5dp"
+                android:layout_width="@dimen/row_item_width_0dp"
+                android:layout_weight="1"
+                android:layout_height="match_parent"/>
+        <Button
+                android:id="@+id/button9"
+                style="?android:attr/buttonBarButtonStyle"
+                android:background="@color/buttonColor"
+                android:layout_margin="@dimen/padding_5dp"
+                android:layout_width="@dimen/row_item_width_0dp"
+                android:layout_weight="1"
+                android:layout_height="match_parent"/>
+    </TableRow>
+
+    <TableRow
+            android:layout_width="match_parent"
+            android:layout_height="@dimen/row_height_0dp"
+            android:layout_weight="1">
+
+        <TextView
+                android:id="@+id/game_result"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center"
+                android:textSize="@dimen/result_view_text_size"
+        />
+
+    </TableRow>
+
+    <Button
+            android:id="@+id/resetButton"
+            android:text="@string/reset"
+            style="?android:attr/buttonBarButtonStyle"
+            android:layout_width="@dimen/row_item_width_0dp"
+            android:layout_height="wrap_content"/>
+
+
+</TableLayout>
