+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
-    <string name="app_name">TickTacToe</string>
+    <string name="app_name">TicTacToe</string>
+    <string name="reset">reset</string>
 </resources>
